backend-v1/requirements.txt
argon2-cffi==20.1.0
bcrypt==3.2.0
black==21.5b1
cryptography==3.4.7
Django==3.2.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
